requirements.txt
fastapi
uvicorn
deepface
minio
python-dotenv
tf-keras
